update to bassantand reco

+++ postgre_database/models.py
@@ -77,7 +77,7 @@ class Rating(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id=Column(Integer, ForeignKey("users.id"))
     place_id=Column(Integer,ForeignKey("places.id")) 
-    rate = Column(Float)
+    rate = Column(Integer)
 
 
 class PlaceType(Base):
